Add files via upload
these are used to calculate profit on 200 tx for any csv of wallet addresses. Logic needs fixed though still a work in progress. Creates a dictionary for every swap transaction that saves the token bought, amount bought, weth spent and then when the sale of these tokens it matches it via token address and amount of tokens swapped (within 90% for tax) and calculates profit based on difference in weth spent/received